+++ src/cli/config/mod.rs
@@ -36,7 +36,6 @@ impl CliContext {
         home: &Option<PathBuf>,
         root: &Option<PathBuf>,
         flag_prof: &Option<RelPathStr>,
-        no_color: bool,
         prompt: Prompt,
     ) -> anyhow::Result<Self> {
         let paths = load_env::load_paths_and_envvars(home, root)?;
@@ -51,11 +50,7 @@ impl CliContext {
         } else {
             curr_profile = root_profile.clone();
         }
-        let col = if no_color || env::var("NO_COLOR").is_ok() {
-            CliColor::nocolor_theme()
-        } else {
-            CliColor::parse_theme(&paths[&Paths::LocalConfigColors])?
-        };
+        let col = CliColor::parse_theme(&paths[&Paths::LocalConfigColors])?;
         Ok(Self {
             paths,
             root_profile,