+++ src/cli/mod.rs
@@ -16,43 +16,39 @@ pub mod prompt;
 #[command(about = "A simple dotfiles manager that doesn't pollute the system", long_about = None)]
 pub struct Cli {
     #[command(subcommand)]
-    cmd: CliCmd,
+    pub cmd: CliCmd,
 
     /// Specify which profile to use
     #[arg(short = 'p', long, env = "AUTOSAVER_PROFILE")]
-    profile: Option<RelPathStr>,
+    pub profile: Option<RelPathStr>,
 
     /// Specify a different home directory to use
     #[arg(short = 'H', long, env = "AUTOSAVER_HOME")]
-    home: Option<PathBuf>,
+    pub home: Option<PathBuf>,
 
     /// Specify a different root directory to use
     #[arg(short = 'R', long, env = "AUTOSAVER_ROOT")]
-    root: Option<PathBuf>,
+    pub root: Option<PathBuf>,
 
     /// Get prompted for each profile before executing it
     #[arg(short = 'c', long)]
-    choice: bool,
-
-    /// Disable all colored output
-    #[arg(short = 'C', long)]
-    no_color: bool,
+    pub choice: bool,
 
     /// Auto answer to all prompts with the specified answers
     #[arg(short = 'a', long, help_heading = "Prompt Options")]
-    auto_answers: Option<String>,
+    pub auto_answers: Option<String>,
 
     /// Auto-answer yes to all prompts
     #[arg(short = 'y', long, conflicts_with_all = ["assume_no", "dry_run"], help_heading = "Prompt Options")]
-    assume_yes: bool,
+    pub assume_yes: bool,
 
     /// Auto-answer no to all prompts
     #[arg(short = 'n', long, conflicts_with_all = ["dry_run", "assume_yes"], help_heading = "Prompt Options")]
-    assume_no: bool,
+    pub assume_no: bool,
 
     /// Skip all prompts and checks entirely
     #[arg(short = 'N', long, conflicts_with_all = ["assume_no", "assume_yes"], help_heading = "Prompt Options")]
-    dry_run: bool,
+    pub dry_run: bool,
 }
 
 #[derive(Subcommand, Debug, Clone, PartialEq, Eq)]
@@ -128,27 +124,27 @@ pub enum CliCmd {
 pub struct CliActSaveRestore {
     /// Allow duplicated paths, and just warn about them
     #[arg(short = 'd', long, global = true)]
-    allow_duplicates: bool,
+    pub allow_duplicates: bool,
 
     /// Allow deleting files in backup directory
     #[arg(short = 'p', long)]
-    allow_purge: bool,
+    pub allow_purge: bool,
 }
 
 #[derive(Args, Debug, Clone, Copy, PartialEq, Eq)]
 pub struct CliActDelSymlinks {
     /// Allow deleting symlink files
     #[arg(short = 's', long, global = true)]
-    allow_symlink: bool,
+    pub allow_symlink: bool,
 }
 
 #[derive(Args, Debug, Clone, Copy, PartialEq, Eq)]
 pub struct CliActBackup {
     /// Show also paths with notdiff policy
     #[arg(short = 'e', long)]
-    show_excluded: bool,
+    pub show_excluded: bool,
 
     /// Show also paths that do not differ
     #[arg(short = 'u', long)]
-    show_unmodified: bool,
+    pub show_unmodified: bool,
 }