+++ src/main.rs
@@ -1,11 +1,18 @@
 use std::process::exit;
 
-use autosaver::{cli::Cli, error};
+use autosaver::{
+    cli::{Cli, inout},
+    error,
+};
 use clap::Parser;
 
 fn main() {
+    // enable colors if not explicitely disabled
+    inout::init_colors();
+
     // parse cmdline
     let cli = Cli::parse();
+
     // run application
     let run_res = cli.run_cmd();